WIP: Start, stop, and supervise gwcelery with systemd
Provide support for starting, stopping, and supervising GWCelery
using systemd, the service manager for Linux. Systemd is present
on many Linux systems including apparently all LDG cluster nodes
and is responsible for bootstrapping userspace and bringing up
system services.

The main advantage of using systemd instead of the existing
htcondor technique is systemd can stop and restart the components
of GWCelery in the proper sequence to minimize downtime and reduce
the chances of losing in-flight tasks.

See the file systemd/README.rst included in this patch for
instructions.