Trigger preliminary alert when preferred event is publishable
Trigger a preliminary alert for a superevent upon the first time
that the preferred event is set to an event that meets the public
alert criterion. Remove the preliminary alert timeout because it
is no longer necessary to wait until the preferred event
stabilizes.

This fixes a longstanding issue that has prevented automated
preliminary alerts from being sent so far. The preferred event
*at the instant that the timeout ended* did not meet the public
alert criterion, but a preferred event that was selected some tens
of seconds later did.